feat: Various large changes
Moved the button press changing code down to be beneath the rest of the
code.
Added in the local mp4 files to demonstrate the issue relating to
the network streaming.
Added in the ability to easily choose whether to
use the network streams or local files at compile time.
Added in play number counter, to see how many plays it takes to break.